virgl: add per-context fencing interface to virgl_context
Being an interface of virgl_context, it is per-context.

This is not to be confused with the existing fencing support.  The
existing fencing support can be considered as ctx0 fencing.  While ctx0
fencing is technically also per-context, it does not use this new
interface because ctx0 is not a virgl_context.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>